Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Why isn't this already in?  Besides it being absolutely ridiculous that a Makefile requires 45 LINES (!!!!) for one object file (it used to require the name of the object file on a SINGLE line which already had a bunch of other object files on it back when I was modifying makefiles in programs of decades past), it is important that this e2label be included, because the current vendor/cyanogen/prebuilt/etc/init.d/05mountsd requires it.

Patch-set: 1