Update patch set 1
Patch Set 1: Code-Review-1

IMO, doing an online resize (a stable feature supported by the kernel) is much cleaner than building a static binary and bundling it with the update.zip.

The failure seems to be because the code to check if the mount point is mounted or not is not working properly. Try http://review.cyanogenmod.org/#/c/79237/. (Make sure you do a clean build or clear the intermediate files for libext2fs for the new flag to take effect.)

Patch-set: 1